Merge branch 'dev' into f/147409-menu-filtro-rol

% Conflicts:
%	src/main/java/es/plexus/hopes/hopesback/configuration/security/WebSecurity.java
%	src/main/java/es/plexus/hopes/hopesback/repository/model/Role.java
%	src/main/java/es/plexus/hopes/hopesback/repository/model/Section.java
%	src/main/resources/postman/Hopes.postman_collection.json
%	src/test/java/es/plexus/hopes/hopesback/controller/UserControllerTest.java
%	src/test/java/es/plexus/hopes/hopesback/service/UserDetailsServiceImplTest.java

+++ src/main/java/es/plexus/hopes/hopesback/repository/model/Role.java
@@ -17,12 +17,12 @@ public class Role {
     @Id
     @Column(name = "rol_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    @Column(name="rol_name", length = 50)
+	@Column(name = "rol_name", length = 50)
     private String name;
 
-    @Column(name="rol_description", length = 500)
-    private String description;
+	@Column(name = "rol_description", length = 500)
+	private String description;
 
 }
